refactor: Move ErrorCallback type alias to base module
