+++ pytrickle/frame_processor.py
@@ -8,7 +8,7 @@ import logging
 from abc import ABC, abstractmethod
 from typing import Optional, Any, Dict, List
 from .frames import VideoFrame, AudioFrame
-from . import ErrorCallback
+from .base import ErrorCallback
 from .state import StreamState, PipelineState
 
 logger = logging.getLogger(__name__)
